requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest-playwright==0.7.0
python-dotenv==1.0.1
allure-pytest==2.14.0
playwright-stealth==1.0.6